Moving signatures to a new ecdsa module
